+++ core/src/main/java/io/snabble/sdk/CheckoutRetryer.java
@@ -130,6 +130,11 @@ class CheckoutRetryer {
                     }
 
                     @Override
+                    public void invalidDepositReturnVoucher() {
+                        fail();
+                    }
+
+                    @Override
                     public void unknownError() {
                         fail();
                     }